+++ src/commands/dev/generate/command.ts
@@ -4,9 +4,8 @@
  * Licensed under the BSD 3-Clause license.
  * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
-import { Flags } from '@oclif/core';
 import { Messages } from '@salesforce/core';
-import { SfCommand } from '@salesforce/sf-plugins-core';
+import { SfCommand, Flags } from '@salesforce/sf-plugins-core';
 import { fileExists, generate } from '../../../util';
 
 Messages.importMessagesDirectory(__dirname);
@@ -31,6 +30,7 @@ export default class GenerateCommand extends SfCommand<void> {
     name: Flags.string({
       required: true,
       summary: messages.getMessage('flags.name.summary'),
+      char: 'n',
     }),
     force: Flags.boolean({
       summary: messages.getMessage('flags.force.summary'),